Moved test data to a new file
